+++ releases/CHANGELOG.md
@@ -1,10 +1,13 @@
 ### auditing-cybersecurity-v1.2.7/Sooo-Under Dev.
 
 Updated: 2023-XX-XX
-Criticality: medium
+Criticality: high
 
 BUG_KNOWN:
 - a lot... too much for describing them without my lawyer present... hhhaha
+
+IMPROVEMENT:
+- '-r | --resources-location': option added (optional) for specifying the path to resources files. Default: '/usr/share/auditing-cybersecurity/resources/'.
 
 OTHERS:
 - g     -> speed response increased.